Renaming MainActivity to LauncherMenuActivity. Cleaning MainActivity class by deleting comments and organizing code. Moving fragment classes to Tab folder. Renaming variables according to convention. Adding tabs to TabLayout from the MainActivity instead of layout file.
