+++ app/src/main/java/com/example/pousadaviladascores/Tabs/Tab1PaginaInicial.java
@@ -1,4 +1,4 @@
-package com.example.pousadaviladascores.Activities;
+package com.example.pousadaviladascores.Tabs;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,9 +9,9 @@ import androidx.fragment.app.Fragment;
 
 import com.example.pousadaviladascores.R;
 
-public class Tab1Pagina_Inicial extends Fragment {
+public class Tab1PaginaInicial extends Fragment {
 
-    public Tab1Pagina_Inicial(){}
+    public Tab1PaginaInicial(){}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
