+++ app/src/main/java/com/example/pousadaviladascores/Tabs/Tab3ItensDeApartamentos.java
@@ -1,4 +1,4 @@
-package com.example.pousadaviladascores.Activities;
+package com.example.pousadaviladascores.Tabs;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,13 +9,13 @@ import androidx.fragment.app.Fragment;
 
 import com.example.pousadaviladascores.R;
 
-public class Tab2Apartamentos extends Fragment {
+public class Tab3ItensDeApartamentos extends Fragment {
 
-    public Tab2Apartamentos(){}
+    public Tab3ItensDeApartamentos(){}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
     {
-        return inflater.inflate(R.layout.tab2_fragment_apartamentos, container, false);
+        return inflater.inflate(R.layout.tab3_fragment_itens_de_apartamentos, container, false);
     }
 }
