+++ app/src/main/res/layout/launcher_menu.xml
@@ -22,30 +22,19 @@
             android:text="@string/pousada_name"
             android:textSize="24sp" />
 
+        <androidx.appcompat.widget.Toolbar
+            android:id="@+id/toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="?attr/colorPrimary"
+            android:minHeight="?attr/actionBarSize"
+            android:theme="?attr/actionBarTheme" />
+
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/tabLayout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-
             app:tabMode="fixed">
-
-            <com.google.android.material.tabs.TabItem
-                android:id="@+id/tab1"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"></com.google.android.material.tabs.TabItem>
-            <!--                android:text="@string/tab_text_1" />-->
-
-            <com.google.android.material.tabs.TabItem
-                android:id="@+id/tab2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"></com.google.android.material.tabs.TabItem>
-            <!--                android:text="@string/tab_text_2" />-->
-
-            <com.google.android.material.tabs.TabItem
-                android:id="@+id/tab3"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"></com.google.android.material.tabs.TabItem>
-            <!--                android:text="@string/tab_text_3" />-->
         </com.google.android.material.tabs.TabLayout>
 
     </com.google.android.material.appbar.AppBarLayout>
