+++ app/src/main/res/layout/tab1_fragment_pagina_inicial.xml
@@ -16,7 +16,7 @@
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
-        android:text="@string/tab_text_1"
+        android:text="@string/tab1_title_Pagina_Inicial"
         android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
