+++ app/src/main/res/layout/tab2_fragment_apartamentos.xml
@@ -16,7 +16,7 @@
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
-        android:text="@string/tab_text_2"
+        android:text="@string/tab2_title_Apartamentos"
         android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
