+++ app/src/main/res/layout/tab3_fragment_itens_de_apartamentos.xml
@@ -16,7 +16,7 @@
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
-        android:text="@string/tab_text_3"
+        android:text="@string/tab3_title_Itens_de_Apartamentos"
         android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
